[Refactor] 리뷰 반영 - soft delete 수행 및 이벤트 발행
